RandLA-Net/requirement.txt
absl-py==0.15.0
argon2-cffi==21.1.0
astor==0.8.1
attrs==22.1.0
backcall==0.2.0
bleach==3.3.1
certifi==2020.6.20
cffi==1.13.0
comm==0.0.1
Cython==0.29.15
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.3
gast==0.6.0
google-pasta==0.2.0
grpcio==1.8.6
h5py==2.10.0
importlib-metadata==2.1.3
ipykernel==5.5.6
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.7.5
jedi==0.17.2
Jinja2==2.11.3
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.6.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.1.3
notebook==6.2.0
numpy==1.16.1
open3d==0.8.0.0
open3d-python==0.3.0.0
packaging==20.9
pandas==0.25.3
pandocfilters==1.5.1
parso==0.7.1
pexpect==4.9.0
pickleshare==0.7.5
prometheus-client==0.12.0
prompt-toolkit==2.0.10
protobuf==3.6.1
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.11.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==5.3
pyzmq==20.0.0
scikit-learn==0.21.3
scipy==1.4.1
Send2Trash==1.8.2
six==1.17.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.6.0
tornado==6.1
traitlets==4.3.3
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.6.10
wrapt==1.15.0
zipp==3.1.0
